Add files via upload
A C program that creates 2 arrays (A and B) of size n x n filled with random numbers, where
n is passed as a command line argument (n >= 128). We computed C, which is the product
of A and B. For a given value of n, we will compute C multiple times, using a different number of threads each
time.